consent and reg debugged, post error with reg
