+++ client/src/components/Consent/consent.css
@@ -1,4 +1,4 @@
-:root {
+/* :root {
     --blue: #7498DA;
     --brown: #593431;
     --pink: #FAF3F0;
@@ -39,4 +39,4 @@
 p {
   font-family: "roboto";
   padding: 40px;
-}
+} */
